requirements.txt
# Teams Meeting Recorder — Python Dependencies
#
# Install with:
#   pip install -r requirements.txt
#
# All packages require Python 3.11+ on macOS (Apple Silicon recommended).

# ── Menu bar UI ──────────────────────────────────────────────────────────────
rumps>=0.4.0                    # macOS menu bar app framework (wraps NSStatusItem)

# ── Audio capture ─────────────────────────────────────────────────────────────
sounddevice>=0.4.6              # Mic + Microsoft Teams Audio virtual device capture
numpy>=1.24.0                   # Audio array manipulation
scipy>=1.11.0                   # WAV file I/O (scipy.io.wavfile)

# ── Transcription: Apple built-in Speech Recognition ──────────────────────────
# No model download needed — uses macOS SFSpeechRecognizer (on-device, offline)
pyobjc-framework-Speech>=10.0   # SFSpeechRecognizer Python bindings

# ── Meeting notes: Claude API ─────────────────────────────────────────────────
anthropic>=0.40.0               # Anthropic Python SDK

# ── Teams process detection ───────────────────────────────────────────────────
psutil>=5.9.0                   # Cross-platform process utilities

# ── macOS system APIs ─────────────────────────────────────────────────────────
pyobjc-core>=10.0               # Python ↔ Objective-C bridge
pyobjc-framework-Cocoa>=10.0    # NSObject, AppKit, Foundation

# ── Configuration ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0            # Load ANTHROPIC_API_KEY from .env file
